--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/sigurn/crc8
+
+go 1.17