- use math/bits reverse
- add go.mod