@tg123 2021-12-31
new
```
go test -bench=CRC8WCDMA -benchtime=20s .
goos: windows
goarch: amd64
pkg: github.com/sigurn/crc8
cpu: Intel(R) Xeon(R) W-2133 CPU @ 3.60GHz
BenchmarkCRC8WCDMA-12           1000000000              10.79 ns/op
PASS
ok      github.com/sigurn/crc8  12.048s
```

old
```
go test -bench=CRC8WCDMA -benchtime=20s .
goos: windows
goarch: amd64
pkg: github.com/sigurn/crc8
cpu: Intel(R) Xeon(R) W-2133 CPU @ 3.60GHz
BenchmarkCRC8WCDMA-12           271214624               88.42 ns/op
PASS
ok      github.com/sigurn/crc8  33.135s
```