@tg123 2021-12-31
after main loop optimize

```
go test -bench=CRC8 -benchtime=20s .
goos: windows
goarch: amd64
pkg: github.com/sigurn/crc8
cpu: Intel(R) Xeon(R) W-2133 CPU @ 3.60GHz
BenchmarkCRC8-12                1000000000               7.859 ns/op
BenchmarkCRC8WCDMA-12           1000000000              10.15 ns/op
PASS
ok      github.com/sigurn/crc8  20.025s
```